Make the call budget a guardrail instead of a fatal error
Reconfigured the five draft n8n workflows, and testing them exposed three real
defects. The Assert Exit Code node added in this batch is what caught all of
them - previously every one would have reported success.

n8n, all five workflows: they used the SSH credential of a different application
(ReportingTeam, not Datawarehouse), pointed at /opt/datawarehouse, piped output
in the shape that discards the exit code, and had no error workflow. opps_sweep
and nightly_reconciliation also passed conflicting sweep windows - the thing that
once made two schedules delete each other's rows.

Enrichment_worker additionally lost data: the SSH node treats any command that
RAN as a success, so a failed run.py flowed straight into "Mark processed" and
retired events that were never enriched. It now asserts first and routes failures
to the deadletter. Its event list is also narrowed to the contract's allowlist -
it included opportunity-created and opportunity-status-changed, ~4,600 events at
one API call each, telling us nothing the sweep and the status ledger give free.

The pipeline defect: exhausting the call budget raised through dlt and aborted
the whole run. Because run.py loops over instances, `ld` spending its 300 calls
meant `local` never ran at all - despite having its own separate quota. With the
wide sweep window, hitting the budget is the NORMAL case while a backlog is being
absorbed, so every scheduled run would have failed and alerted. An alert that
always fires is an alert nobody reads.

Budget exhaustion now stops enrichment cleanly, lands what was fetched, and
resumes from the same watermark next run. Both the transformer and the sweep
generator catch it: dlt pulls them round-robin, so once the transformer has spent
the budget it is the sweep's NEXT page that raises - catching it in one place left
it escaping from the other. A truncated sweep still refuses to record itself as
complete, so the deletion pass keeps using the last full sweep rather than
concluding that everything it never reached has vanished.

Also declares the opportunity_deletions source. The model guarded on the relation
existing but read it through source(), which needs a declaration too - so the
build broke the first time a deletion was ever detected, having been green for
weeks. 550 deletions landed; none appears in the report, which is the independent
check available.

Verified end to end from n8n: status success, ok true, both instances LOADED with
no failed jobs. Droplet PASS=204 ERROR=0. Coverage: money 2,679 -> 2,971
opportunities, $8.62M -> $10.06M.

+++ dbt/models/staging/stg_smartmoving__opportunities_enriched.sql
@@ -20,6 +20,11 @@
 --
 -- pipeline_status arrives with inconsistent trailing whitespace - both 'Booked'
 -- and 'Booked ' occur - so it is trimmed once here at the boundary.
+--
+-- MONEY IS CAST TO numeric HERE. dlt lands estimated_total__* as double precision
+-- because that is the JSON shape, and binary floats drift by fractions of a cent
+-- under SUM. Typing is staging's job (CLAUDE.md: "renamed, typed, lightly
+-- cleaned"), so the cast lives here once instead of in every downstream model.
 
 select
     source_instance_id || ':' || id             as opportunity_key,
@@ -53,10 +58,10 @@ select
     nullif(tariff__id, '')                      as tariff_id,
     nullif(tariff__name, '')                    as tariff_name,
 
-    estimated_total__subtotal                   as estimated_subtotal,
-    estimated_total__taxable_amount             as estimated_taxable_amount,
-    estimated_total__tax                        as estimated_tax,
-    estimated_total__final_total                as estimated_final_total,
+    estimated_total__subtotal::numeric          as estimated_subtotal,
+    estimated_total__taxable_amount::numeric    as estimated_taxable_amount,
+    estimated_total__tax::numeric               as estimated_tax,
+    estimated_total__final_total::numeric       as estimated_final_total,
 
     nullif(move_size__name, '')                 as move_size_name,
     nullif(move_size__description, '')          as move_size_description,
